hotfix: 비로그인시 search 화면 렌더링 안되는 문제 + 라우팅 버그 수정 (#355)
