+++ src/app/review/write/_component/DirectMyPetInfo.tsx
@@ -72,6 +72,7 @@ const DirectMyPetInfo = ({
         {/* 1-3-1. 종 */}
         <div className={styles.halfTextField}>
           <span>종</span>
+          {/* tavian */}
           <TextField
             value={petType}
             onClick={() => setActiveDropDown((prev) => (prev === "petType" ? null : "petType"))}
@@ -102,6 +103,7 @@ const DirectMyPetInfo = ({
             control={control}
             render={({ field }) => (
               <>
+                {/* tavian */}
                 <TextField
                   value={field.value === "F" ? "암컷" : field.value === "M" ? "수컷" : ""}
                   onClick={() => setActiveDropDown((prev) => (prev === "gender" ? null : "gender"))}
